requirements.txt
alembic==1.11.1
anyio==3.7.0
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
defusedxml==0.7.1
Django==4.2.2
django-cors-headers==4.1.0
django-jazzmin==2.6.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-yasg==1.21.6
exceptiongroup==1.1.1
fastapi==0.98.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
inflection==0.5.1
Mako==1.2.4
MarkupSafe==2.1.3
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
psycopg2==2.9.6
pycparser==2.21
pydantic==1.10.9
PyJWT==2.7.0
pyTelegramBotAPI==4.12.0
python-decouple==3.8
python-dotenv==1.0.0
python-magic==0.4.27
python-magic-bin==0.4.14
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
SQLAlchemy==2.0.17
sqlalchemy-media==0.17.4
sqlparse==0.4.4
starlette==0.27.0
telebot==0.0.5
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
uvicorn==0.22.0
Wand==0.6.11

sqlitis~=0.0.5